Split up bokeh.models.renderers and move contour_renderer
