--- docs/bokeh/source/docs/reference/models/contour_renderer.rst
@@ -1,8 +1,0 @@
-.. _bokeh.models.contour_renderer:
-
-contour renderer
-----------------
-
-.. automodule:: bokeh.models.contour_renderer
-   :members:
-   :undoc-members:
