+++ src/bokeh/models/renderers/__init__.py
@@ -1,0 +1,62 @@
+#-----------------------------------------------------------------------------
+# Copyright (c) 2012 - 2022, Anaconda, Inc., and Bokeh Contributors.
+# All rights reserved.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+""" Various kinds of renderers.
+
+"""
+#-----------------------------------------------------------------------------
+# Boilerplate
+#-----------------------------------------------------------------------------
+from __future__ import annotations
+
+import logging # isort:skip
+log = logging.getLogger(__name__)
+
+#-----------------------------------------------------------------------------
+# Imports
+#-----------------------------------------------------------------------------
+
+# Bokeh imports
+from . import (
+    contour_renderer,
+    glyph_renderer,
+    graph_renderer,
+    renderer,
+    tile_renderer,
+)
+from .contour_renderer import *
+from .glyph_renderer import *
+from .graph_renderer import *
+from .renderer import *
+from .tile_renderer import *
+
+#-----------------------------------------------------------------------------
+# Globals and constants
+#-----------------------------------------------------------------------------
+
+__all__ = (
+    *contour_renderer.__all__,
+    *glyph_renderer.__all__,
+    *graph_renderer.__all__,
+    *renderer.__all__,
+    *tile_renderer.__all__,
+)
+
+#-----------------------------------------------------------------------------
+# General API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Dev API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Private API
+#-----------------------------------------------------------------------------
+
+#-----------------------------------------------------------------------------
+# Code
+#-----------------------------------------------------------------------------
